requirements.txt
Pillow
pyright
ruff
pre-commit